Rename type_generation module to gen
